# apidefinition/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Meeting app
  description: Basic meeting reservation api
  version: 1.0.9

servers:
  - url: http://localhost:8080/api/
    description: Develop

tags:
  - name: meeting

paths:
  /private/meeting/findbytime/{datetime}:
    get:
      tags:
        - MeetingReservation
      summary: Returns meeting reservation by datetime
      operationId: findByDateTime
      parameters:
        - name: datetime
          in: path
          description: search parameter in format yy.MM.dd hh:mm
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Response when meeting is reserved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingReservationDto"
        "4XX":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "5XX":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"

  /private/meeting/listall:
    get:
      tags:
        - MeetingReservation
      summary: List all meetings
      operationId: listAll
      responses:
        "200":
          description: Response when meeting is reserved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingListDto"

  /private/meeting/deleteall:
    get:
      tags:
        - MeetingReservation
      summary: deletes all meetings
      operationId: deleteAll
      responses:
        "200":
          description: ok

  /private/meeting/reservation:
    post:
      tags:
        - MeetingReservation
      summary: Post meeting reservation data
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingReservationRequestDto"
      responses:
        "200":
          description: Response when meeting is reserved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingReservationDto"
        "4XX":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "5XX":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"

components:
  schemas:
    MeetingListDto:
      type: array
      items:
        $ref: "#/components/schemas/MeetingReservationDto"

    MeetingReservationDto:
      type: object
      properties:
        day:
          type: string
          format: date
        start:
          type: string
        end:
          type: string
        user:
          type: string

    MeetingReservationRequestDto:
      type: object
      required:
        - username
        - day
        - startHour
        - startMin
        - endHour
        - endMin
      properties:
        username:
          type: string
          example: User Ubul
        day:
          type: string
          format: date
          description: "reservations day, format: yyyy-MM-dd"
        startHour:
          type: integer
          maximum: 16
          minimum: 9
          description: "starting hour of meeting, min 9, max 16"
        endHour:
          type: integer
          minimum: 9
          maximum: 17
          description: "ending hour of meeting, min 9, max 17"
        startMin:
          type: integer
          minimum: 0
          maximum: 59
          description: "starting minute of meeting, must be 0 or 30"
          example: 30
        endMin:
          type: integer
          minimum: 0
          maximum: 59
          description: "ending minute of meeting, must be 0 or 30"
          example: 0

    ErrorDto:
      type: object
      properties:
        message:
          type: string
        errorList:
          type: array
          items:
            type: object
            properties:
              message:
                type: string